+++ lib/friendly_id/reserved.rb
@@ -7,21 +7,15 @@ module FriendlyId
 The {FriendlyId::Reserved Reserved} module adds the ability to exlude a list of
 words from use as FriendlyId slugs.
 
-By default, FriendlyId reserves the words "new" and "edit" when this module is
-included. You can configure this globally by using {FriendlyId.defaults
-FriendlyId.defaults}:
+With Ruby on Rails, FriendlyId's generator generates an initializer that
+reserves the words "new" and "edit" using {FriendlyId.defaults
+FriendlyId.defaults}.
 
-  FriendlyId.defaults do |config|
-    config.use :reserved
-    # Reserve words for English and Spanish URLs
-    config.reserved_words = %w(new edit nueva nuevo editar)
-  end
-
-Note that the error message will appear on the field `:friendly_id`. If you are
-using Rails's scaffolded form errors display, then it will have no field to
-highlight. If you'd like to change this so that scaffolding works as expected,
-one way to accomplish this is to move the error message to a different field.
-For example:
+Note that the error messages for fields will appear on the field
+`:friendly_id`. If you are using Rails's scaffolded form errors display, then
+it will have no field to highlight. If you'd like to change this so that
+scaffolding works as expected, one way to accomplish this is to move the error
+message to a different field. For example:
 
   class Person < ActiveRecord::Base
     extend FriendlyId
@@ -42,7 +36,7 @@ For example:
     def self.included(model_class)
       model_class.class_eval do
         friendly_id_config.class.send :include, Reserved::Configuration
-        friendly_id_config.defaults[:reserved_words] ||= ["new", "edit"]
+        friendly_id_config.defaults[:reserved_words] ||= []
       end
     end