+++ Changelog.md
@@ -5,6 +5,7 @@ suggestions, ideas and improvements to FriendlyId.
 
 ## 5.0.0.beta4 (NOT RELEASED YET)
 
+* Add an initializer to the generator; move the default reserved words there.
 * Allow assignment from {FriendlyId::Configuration#base}.
 
 ## 5.0.0.beta3 (2013-08-20)